requirements.txt
aws-cdk.core==1.38.0
aws-cdk.aws-iam==1.38.0
aws-cdk.aws-logs==1.38.0
aws-cdk.aws-lambda==1.38.0
aws-cdk.aws-events==1.38.0
aws-cdk.aws-events-targets==1.38.0
boto3==1.13.10